+++ scout/adapter/mongo/institute.py
@@ -49,6 +49,7 @@ class InstituteHandler(object):
         add_groups=None,
         sharing_institutes=None,
         cohorts=None,
+        loqusdb_ids=[],
         alamut_key=None,
     ):
         """Update the information for an institute
@@ -68,13 +69,13 @@ class InstituteHandler(object):
             add_groups(bool): If groups should be added. If False replace groups
             sharing_institutes(list(str)): Other institutes to share cases with
             cohorts(list(str)): patient cohorts
+            loqusdb_ids(list(str)): list of ids of loqusdb instances Scout is connected to
             alamut_key(str): optional, Alamut Plus API key -> https://extranet.interactive-biosoftware.com/alamut-visual-plus_API.html
 
         Returns:
             updated_institute(dict)
 
         """
-
         add_groups = add_groups or False
         institute_obj = self.institute(internal_id)
         if not institute_obj:
@@ -155,9 +156,9 @@ class InstituteHandler(object):
         if cohorts is not None:
             updates["$set"]["cohorts"] = cohorts
 
-        if loqusdb_id is not None:
-            LOG.info("Updating loqusdb id for institute: %s to %s", internal_id, loqusdb_id)
-            updates["$set"]["loqusdb_id"] = loqusdb_id
+        if loqusdb_ids:
+            LOG.warning("Updating loqusdb id for institute: %s to %s", internal_id, loqusdb_ids)
+            updates["$set"]["loqusdb_id"] = loqusdb_ids
 
         if alamut_key is not None:
             updates["$set"]["alamut_key"] = (